version 1.74
Finally fixed the address of NArw call for categorical sets, such that now the NArw call is addressed separately in the transform_dict as a grandparent primitive consistent with the other categories. Also cleaned up the code to resolve all of the warning messages that were being returned by scikit and pandas, those were kind of annoying. Finally just some cleanup for the formatting of documentation for function definitions.