v2.17
updated description of df_train and df_test sets to include note:
"The tool supports the inclusion of non-index-range column as index or 
multicolumn index (requires named index columns). Such index types 
are added to the returned "ID" sets which are consistently shuffled and 
partitioned as the train and test sets.